Update MailWrap for Mail 9.0 and OS X 10.11 'El Capitan'
The DocumentEditor class has been replaced by ComposeViewController and
the cursor now needs to be positioned in the show selector instead of
finishLoadingEditor. Rework the attribution clean-up code to fit this
brave new world.

The version of py2app shipping with OS X 10.11 is broken by the new
System Integrity Protection feature. Document this with instructions for
disabling 'rootless' mode and removing the offending file flags until
the system py2app is properly fixed.