Unconditionally fix attribution quote level
The FixAttribution default toggles whether the 'On DATE, at TIME, '
prefix is stripped from attribution lines, but we need to fix the
broken quote level in both cases.

The Apple Mail bug which leads to attribution lines being incorrectly
quoted is now nearly a decade old. Perhaps we're holding it wrong?